requirements.txt
django==4.1.4
djangorestframework==3.12.1
psycopg2==2.9.5
pytz